broke cratedb types out class-per-file style
